Move `testCreateLinkPreviewsOption_TC_25795` to SettingsTests
