+++ wire-ios/WireUITests/ConversationTests.swift
@@ -107,57 +107,4 @@ final class ConversationTests: WireUITestCase {
 
         XCTAssertFalse(activeConversationPage.userLeftSystemMessage.exists, "the system message has not been removed")
     }
-
-    @MainActor
-    func testCreateLinkPreviewsOption_TC_25795() async throws {
-        let stagingTeam = try await UserHelper.default.registerTeam(withMemberCount: 2)
-        let userA = try XCTUnwrap(stagingTeam.teamMembers.first)
-        let userB = try XCTUnwrap(stagingTeam.teamMembers.last)
-
-        // Login & create conversation
-        _ = try await loginToBackend(user: userA)
-            .tapPlusButtonToCreateGroup()
-            .tapNewGroupButton()
-            .enterGroupName("disabled preview test")
-            .tapMemberCells(withLabelPrefixes: [userB.name])
-            .doneSelectingMembers()
-            .goBackToConversationPage()
-
-        // Disable link previews
-        _ = try ConversationsPage()
-            .openSettings()
-            .openOptionsMenu()
-            .disableCreateLinkPreviews()
-            .backToSettings()
-            .switchToConversationsTab()
-
-        // Open conversation and send first link
-        _ = try ConversationsPage()
-            .openConversation()
-            .sendMessage("First link: https://github.com/wireapp/wire-ios")
-            .goBackToConversationPage()
-
-        // Enable link previews
-        _ = try ConversationsPage()
-            .openSettings()
-            .openOptionsMenu()
-            .enableCreateLinkPreviews()
-            .backToSettings()
-            .switchToConversationsTab()
-
-        // Open conversation and send first link
-        let conversationPage = try ConversationsPage()
-            .openConversation()
-            .sendMessage("Second link: https://github.com/wireapp/wire-android")
-
-        // Verify first message is the original message without preview
-        _ = conversationPage
-            .verifyMessageSent("First link: https://github.com/wireapp/wire-ios")
-
-        // Verify second message has link preview
-        _ = conversationPage
-            .verifyMessageSent("Second link:")
-            .verifyLinkPreviewCell()
-    }
-
 }
