+++ wire-ios/WireUITests/SettingsTests.swift
@@ -1,0 +1,74 @@
+//
+// Wire
+// Copyright (C) 2026 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import XCTest
+
+final class SettingsTests: WireUITestCase {
+
+    @MainActor
+    func testCreateLinkPreviewsOption_TC_25795() async throws {
+        let stagingTeam = try await UserHelper.default.registerTeam(withMemberCount: 2)
+        let userA = try XCTUnwrap(stagingTeam.teamMembers.first)
+        let userB = try XCTUnwrap(stagingTeam.teamMembers.last)
+
+        // Login & create conversation
+        _ = try await loginToBackend(user: userA)
+            .tapPlusButtonToCreateGroup()
+            .tapNewGroupButton()
+            .enterGroupName("disabled preview test")
+            .tapMemberCells(withLabelPrefixes: [userB.name])
+            .doneSelectingMembers()
+            .goBackToConversationPage()
+
+        // Disable link previews
+        _ = try ConversationsPage()
+            .openSettings()
+            .openOptionsMenu()
+            .disableCreateLinkPreviews()
+            .backToSettings()
+            .switchToConversationsTab()
+
+        // Open conversation and send first link
+        _ = try ConversationsPage()
+            .openConversation()
+            .sendMessage("First link: https://github.com/wireapp/wire-ios")
+            .goBackToConversationPage()
+
+        // Enable link previews
+        _ = try ConversationsPage()
+            .openSettings()
+            .openOptionsMenu()
+            .enableCreateLinkPreviews()
+            .backToSettings()
+            .switchToConversationsTab()
+
+        // Open conversation and send first link
+        let conversationPage = try ConversationsPage()
+            .openConversation()
+            .sendMessage("Second link: https://github.com/wireapp/wire-android")
+
+        // Verify first message is the original message without preview
+        _ = conversationPage
+            .verifyMessageSent("First link: https://github.com/wireapp/wire-ios")
+
+        // Verify second message has link preview
+        _ = conversationPage
+            .verifyMessageSent("Second link:")
+            .verifyLinkPreviewCell()
+    }
+}
